requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.4.0
attrs==25.3.0
azure-core==1.34.0
azure-identity==1.22.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.3
cryptography==44.0.3
frozenlist==1.7.0
idna==3.10
msal==1.34.0
msal-extensions==1.3.1
multidict==6.6.4
numpy==2.2.5
propcache==0.3.2
pycparser==2.22
PyJWT==2.10.1
python-dotenv==1.1.0
requests==2.32.3
six==1.17.0
sounddevice==0.5.1
typing_extensions==4.13.2
urllib3==2.4.0
websocket-client==1.8.0
yarl==1.20.1
